chatbot-agno/requirements.txt

agno
openai
streamlit
python-dotenv
